[diem-framework] specs for NetworkIdentity
This commit adds the specifications for the `NetworkIdentity` module and
changes a few implementation pieces in order to keep the SMT solver
happy.

The `NetworkIdentity` should be fully specified now. An interesting part
in this code is to simulate two set operations, set union and set
difference, via vector operations. For this, two generic functions are
added, `add_members_internal` and `remove_members_internal` and they
might be useful outside the context of `NetworkIdentity` as well.